refactor(llama): split loader out
