+++ llama-rs/src/util.rs
@@ -2,6 +2,21 @@ use std::path::{Path, PathBuf};
 
 use crate::LoadError;
 
+/// NOTE: The original code relies in promotion rules and automatic cast between
+/// int to float. What we do instead is use this macro to convert every term of
+/// the multiplication to f64, which should have enough precision bits to hold
+/// the final value, then cast to usize. I have observed a discrepancy between
+/// the ctx_size found using this code, and the one in llama.cpp. The number for
+/// rust ends up being slightly lower, but no "out of memory" errors are
+/// reported by ggml.
+macro_rules! mulf {
+    ($term:expr, $($terms:expr),*) => {
+        usize::try_from((($term as f64) $(* ($terms as f64))*) as u64).unwrap()
+    };
+}
+
+pub(crate) use mulf;
+
 /// Used to buffer incoming tokens until they produce a valid string of UTF-8 text.
 ///
 /// Tokens are *not* valid UTF-8 by themselves. However, the LLM will produce valid UTF-8
